fix(consensus): improve block fullness when tx rate is high (#916)
Description
---
Remove calls to `beat` when receiving a local proposal
Only call `beat` on new transactions when there no transactions pending
execution
Reduced some info log noisiness

Motivation and Context
---
`beat` was called when receiving a local proposal, however this is not
necessary unless you are the leader. `beat` is called when receiving
votes (including from self) as the next leader so this PR essentially
only calls beat on reciept of a proposal if it is the next leader.

If there is only a single transaction, we avoid waiting for the block
time by allowing the mempool to send how many transactions are awaiting
execution allowing consensus to process the single transaction
immediately, or wait for multiple transactions to finish executing
before proposing.

How Has This Been Tested?
---
Manually, stess test and single transactions.

250 stress test with 2 shards:

![image](https://github.com/tari-project/tari-dan/assets/1057902/b614aedc-4633-4463-bd09-f3638f6da80d)

What process can a PR reviewer use to test or verify this change?
---
Stress test and check how many commands are proposed in a block

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify