fix(engine)!: determinism bug with log entries (#926)
Description
---
Remove timestamp from LogEntry to fix determinism bug in txreceipt
Removed timestamp from UI

Motivation and Context
---
Initially logs were not part of any substate, but these are added in
TxReceipt. Logs contained a timestamp based on the local clock of the
VN.

This PR removes the timestamp, ordering is maintained by the Vec.

How Has This Been Tested?
---
Not explicitly tested. This caused Merkle root mismatches in #924 

What process can a PR reviewer use to test or verify this change?
---
Check that tx receipt does not have local time in it

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify