feat: transparent bindings and bigints (#919)
Description
---
Handle the `#[serde(transparent)]` structs, and also fix the bigint. i64
and u64 were exported as bigint, instead of number. The number can hold
anything in range(inclusive) from -(2^53-1)...(2^53-1). Anything above
or below can be hold but it's not safe to be precise.

Motivation and Context
---
Some of the structures are `#[serde(transparent)]` this is not
recognized by the ts-rs crate. If the structure is "real struct" (e.g.
struct A{b:u64} and not a struct A(u64), then we can call #[ts(flatten)]
but you have to call it on every instance. And it will rewrite the
variable e.g.
```
struct A{
  #[ts(flatten)]
  b:B
}

struct B {
  c:u8
}
```
Will be exported as
```
export interface A {
  c : number; // instead of b : number
}
```
So I have decided instead of flatten define the types by hand 

How Has This Been Tested?
---

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify