fix(walletd): emit AccountChanged event on changes for account refresh (#956)
Description
---
Emit missing event if the account is updated via refresh.

Motivation and Context
---
The account monitor only emitted the AccountChanged event after a
successful transaction involving the account is detected. However, if a
client (e.g. cucumber) is polling the account refresh, the refresh can
detect changes to the account before the transaction is detected as
complete. This is the typical case as the cucumber polls faster than the
wallet polls the indexer for a transaction result.

This bug led to `Claim and transfer confidential assets via wallet
daemon` never finishing because once the finalised transaction comes
through, there are no changes detected because the refresh already
updated the account.

How Has This Been Tested?
---
`Claim and transfer confidential assets via wallet daemon` cucumber
completes.

What process can a PR reviewer use to test or verify this change?
---
CI

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify