fix!: always account for version when querying substates (#966)
Description
---
Hash in the version parameter when generating non-versionable substates
Fix race condition when waiting for transaction and account results

Motivation and Context
---
Although the unspent commitment, fee claim, and tx receipt are
non-versionable. We cannot ignore the version parameter because clients
may query these substates with a version > 0, in this case the VNs
should respond with `DoesNotExist`. Previously, due to ignoring the
version, we would respond either with UP or DOWN if the substate exists
`@v0` regardless of the version. For a DOWN result, the scanner would
increment the version and query again, leading to an infinite loop.

How Has This Been Tested?
---
`Double Claim base layer burn funds with wallet daemon. should fail`
cucumber passes

What process can a PR reviewer use to test or verify this change?
---
CI, query a spent commitment, fee claim or tx receipt with v > 0 and
check that DoesNotExist is returned


Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [x] Other - Please specify

BREAKING CHANGE: substate address for unspent commitment, fee claim, and
tx receipt are derived differently