feat: add block hash (#965)
Description
---
Add base layer block hash to the block. This only adds the block, does
nothing with it yet.

Motivation and Context
---
This is going to be used in the upcoming changes for the epoch
management.

How Has This Been Tested?
---
See below.

What process can a PR reviewer use to test or verify this change?
---
Start the network, and look in the DB blocks table and compare with
data.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify