refactor(tx-builder): allow building unsigned transactions (#939)
Description
---
refactor(tx-builder): allow building unsigned transactions

Motivation and Context
---
We have a few interfaces for submitting transactions, each of these has
several fields that are duplicated across types. Some have missing
fields simply because a particular transaction feature is unused. This
PR adds the UnsignedTransaction type that is built using the
TransactionBuilder and given to a wallet for signing.

We could also build a typescript version of the transaction builder to
make building well-formed transactions easy.

This PR also adds an optional field to accept the UnsignedTransaction to
the wallet daemon. If not specified, the separate fields are used as
before to preserve backward compatibility. The idea would be to remove
these later in another PR.

How Has This Been Tested?
---
Updated some cucumber steps to use this

What process can a PR reviewer use to test or verify this change?
---
```rust
   let transaction = Transaction::builder()
        .fee_transaction_pay_from_component(source_component_address, Amount(1))
        .call_method(source_component_address, "withdraw_confidential", args![
            resource_address,
            withdraw_proof
        ])
        .put_last_instruction_output_on_workspace(b"bucket")
        .call_method(destination_account, "deposit", args![Workspace("bucket")])
        .build_unsigned_transaction();

// Send off to the wallet for signing and submission
    let submit_req = TransactionSubmitRequest {
        transaction: Some(transaction),
       signing_key_index: Some(signing_key_index),
     ..Default::default()
   }
```

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify