refactor(engine)!: store NFT data as CBOR value (#957)
Description
---
refactor(engine)!: store NFT data as CBOR value

Motivation and Context
---
Internally represent the `data` and `mutable_data` fields of NFTs as a
tari_bor::Value (ciborium). This reduces the decoding and encoding
required to operate on the data and brings NFTs inline with other code
(e.g. component state).

This also allows for a JSON representation in client requests e.g. get
substates, that can be interrogated easily without decoding an opaque
blob first. For example, there is no library currently for decoding CBOR
in the browser. This JSON representation can be used to easily extract
data e.g. user data in an NFT.

How Has This Been Tested?
---
Use the stablecoin test website, substates.get JRPC call in the wallet
returns a JSON representation instead of opaque bytes.

```json
  "NonFungible": {
        "data": {
          "created_at": 0,
          "user_id": 1
        },
        "mutable_data": {
          "is_blacklisted": false,
          "wrapped_exchange_limit": 1000
        }
      }
    }
```

What process can a PR reviewer use to test or verify this change?
---
Call substates.get, or look in the database

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [x] Other - Please specify

BRAKING CHANGES: due to the changes in the template library to encode as
a tari_bor::Value, templates that use NFT data will need to be
recompiled. NFT data in the persistent substate store is not backward
compatible.