fix(consensus)!: add exhaust burn per transaction, fix block fee malleability (#942)
Description
---
~~fix(consensus)!: calculate fees per block~~
fix(consensus)!: immalleable block fee 
feat(consensus)!: enable fixed exhaust of 5%

Motivation and Context
---
~~Previously, fees and exhaust burns were calculated per transaction.
This increases the deviation caused by rounding and isa little more
complex implementation-wise. This PR sums up fees for a block and then
computes the leader fee and exhaust burn.~~

The leader fee calculation is improved to better account for burnt fees
due to the remainder when dividing the fees by the number of involved
shards.

This PR enables the exhaust at a fixed rate of 5% - this number is not
finalized and will change but is enabled to allow for better testing.

The fee field in the block was previously malleable, this PR includes
the fees in the block hash, and therefore block signature.

How Has This Been Tested?
---
New unit test, manually checked that fee agreement is maintained

What process can a PR reviewer use to test or verify this change?
---
Check the expected fee and burn rate in VN database

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify