feat: store base layer block hashes (#968)
Description
---
Store the base layer block hashes (not just the ones we sync, but to the
very tip of the chain).

Motivation and Context
---
We will need this for accepting/rejecting the proposed blocks. Currently
we have base layer block hash in the blocks, but we do nothing with it.
We need to validate if the hash is indeed correct. For that we need all
the hashes from the base layer.

What process can a PR reviewer use to test or verify this change?
---
Look into the `global_storage.sqlite` table to see that the correct data
is stored.


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify