chore(integration_test): show more info about wallet daemons (#974)
Description
---
Added port and path info about wallet daemons in cucumber

Motivation and Context
---
In cucumber, when using the `When I print the cucumber world` step, it
only shows the wallet daemon name. This makes it difficult for manual
testing if we want to stop the execution and inspect the wallet daemon.

This PRs adds port and path info similar to VNs and indexers.

How Has This Been Tested?
---
Inspecting the result of the `When I print the cucumber world` cucumber
step

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify