feat(indexer): new GraphQL query to get events by payload (#985)
Description
---
* New GraphQL `getEventsByPayload` query, that returns (allowing
pagination) all the events that contain a specific key-value in their
payloads
* Updated SQLite database with a new table `event_payloads` to store all
key-value of the payloads, referencing the already existing `events`
table.
* For efficiency, the `events` table still maintains a `payload` field
containing the JSON value. This avoids a JOIN operation on the payload
key-values, and there is no inconsistency because events are read-only.
The payload key-values are used just for filtering events in queries.
* To be able to insert into the new one-to-many `event_payloads` table,
we need the `returning_clauses_for_sqlite_3_35` feature for `diesel` so
we can get the `id` of a newly created `event` row.
* Added all the needed logic into managers and DB transaction wrappers
to allow to save the payloads into the new table as well as to fetch all
events that have some specific key-value

Motivation and Context
---
We want to be able to query indexers for events that include a
particular payload key-value, for example to fetch the new builtin vault
events (https://github.com/tari-project/tari-dan/pull/975) by resource
address.

The indexer already has a GraphQL API to get events based on other
criteria (by components and by templates), and this PR adds a new method
to that api that:
* Allows to get all events that include a specific key-value in their
`payload` field
* Allows pagination so the client can specify the `offset` and `limit`
of the results

This PR **does not implement** the Indexer-VN logic needed to scan those
events, it only implements the GraphQL and database operations for
storing/fetching events by payload.

How Has This Been Tested?
---
New integration test for the new GraphQL query. As the Indexer is still
not able to scan by payload from VNS, I tested by inserting some mock
events on indexer startup so the new test fetches them.

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>