fix(wallet): resubmit transactions on failure (#978)
Description
---
fix(wallet): resubmit status=New transactions on failure
refactor(wallet): remove duplicate fee receipt

Motivation and Context
---
If a wallet fails to submit a transaction e.g. due to a network failure,
it previously is not resubmitted and stays as a status = New
transaction. This PR attempts to resubmit New transactions. This does
not apply to failed transactions but simply to transactions that never
reached the network.

How Has This Been Tested?
---
Manually by shutting down the VN/indexer, submitting a transaction, and
restarting the VN or indexer.

What process can a PR reviewer use to test or verify this change?
---
As above

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify