feat: store substate_id in events (#986)
Description
---
* Updated the `events` database table for the new field
* Updated database API and managers to use substate ids instead of
component addresses
* Builtin vault events (e.g. `std.vault.deposit`) now emit two events:
one has the vault as `substate_id` and the other has the resource. This
is done to allow queries by both criteria

Motivation and Context
---
Up until now events stored the component address related to it. This is
not general enough now because we want also to emit events that can be
related to other type of substates (e.g. resources or vaults). This PR
changes the `component_address` field for `substate_id` for events in
the whole project (from database to APIs)

How Has This Been Tested?
---
* Existing unit and integration tests pass.
* Manually inspected the content of the new `substate_id` field in
different types of events

What process can a PR reviewer use to test or verify this change?
---
Emit any type of event and check the new `substate_id` field value

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Event fields change (`substate_id` instead of `component_address`)
so it requires network reset

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>