feat!: check the incoming base layer block hash in proposals (#972)
Description
---
We are adding the check for the base layer hash in the proposal to be
within some +- from the current synced height. And it has to be known.
The check if the proposed block is too much behind is disabled. It's
causing issues during sync (added TODO).

Motivation and Context
---

How Has This Been Tested?
---
I just did some logging to see if the height is being compared. But not
the failing path.

What process can a PR reviewer use to test or verify this change?
---
Maybe run a couple vns, with running miner and see what's going on, will
it fail? Make sure you run multiple committees to also test that
foreignproposals will be validated.

Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - The 2nd layer should be restarted, otherwise the blocks
validation will fail. But wait for the next PR that also have new
breaking changes.