chore(integration_test): show GraphQL port in cucumber world (#1009)
Description
---
Added indexer's GraphQL port info in cucumber

Motivation and Context
---
In cucumber, when using the `When I print the cucumber world` step, it
does not show the indexer's GraphQL port. This makes it difficult for
manual testing if we want to stop the execution and query the indexer's
GraphQL API for events.

This PRs adds the GraphQL port info to the cucumber world display.

How Has This Been Tested?
---
Inspecting the result of the `When I print the cucumber world` cucumber
step

What process can a PR reviewer use to test or verify this change?
---
See previous section

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify