fix: use fallback jrpc address in webui for indexer and wallet (#1012)
Description
---
fix: use fallback jrpc address in webui for indexer and wallet

Motivation and Context
---
When we're unable to bind on the preferred JRPC port, an OS-selected
port is used. This PR ensures that the webui uses the correct port in
this case.

This was already implemented on the validator node, but not on the
indexer and wallet

How Has This Been Tested?
---
Manually

What process can a PR reviewer use to test or verify this change?
---
Open multiple indexers/wallets and check that their webuis use the
correct JRPC address.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify