fix(p2p): configurable relay limits, set noise prologue for relay client (#1017)
Description
---
configurable relay limits
set noise prologue for relay client

Motivation and Context
---
Some nodes will act primarily as relays, this increases the relay
resource limits to accommodate this.

@stringhandler found a potential issue where the relay client was not
using the same noise prologue as the base swarm instance. This was an
oversight and corrected in this PR.

How Has This Been Tested?
---
In progress

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify