feat!: implements epoch change (#982)
Description
---
Add epoch events
- `End` this is added at the last base layer block that is still in the
epoch. We have to bury the block to be locked, after that we officially
end the epoch
- `Start` this will be in the first block when the epoch switch, its
parent will have `End` or it will be genesis block.

Motivation and Context
---

How Has This Been Tested?
---
I've tested multiple committees and I switched epoch many times. I've
checked the blocks and the commands in them to confirm the epoch
end/start.
There is also a cucumber test. That checks if the epoch ended with a
EpochEnd command and new one started with EpochStart command.

What process can a PR reviewer use to test or verify this change?
---
Do as above, or go a step further and try to make a transaction during
epoch switching when you also add a VNs that changes space division.


Breaking Changes
---

- [ ] None
- [ ] Requires data directory to be deleted
- [x] Other - The 2nd layer should be restarted, otherwise the blocks
validation will fail.

---------

Co-authored-by: Stan Bondi <sdbondi@users.noreply.github.com>