fix(engine): correctly revert to fee checkpoint if insufficient fees paid (#1018)
Description
---
revert to fee checkpoint if insufficient fees paid after successful main
transaction execution.

Motivation and Context
---
If the user paid sufficient fees for the fee transaction and
insufficient fees for the main transaction, the transaction processor
would still return a substate diff for the changes in the main
transaction. This PR fixes this.

How Has This Been Tested?
---
Updated existing test for this case to explicitly check that this does
not occur (failing test). This test passes after these changes.

What process can a PR reviewer use to test or verify this change?
---
Pay sufficient fees for the fee transaction, but insufficient for the
main transaction.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify