refactor!: fix epoch database structure (#1034)
Description
---
Changes the epoch, validator and committee structure in the global
database.


Motivation and Context
---
Previously, validator nodes did not deregister and this resulted in a
bug that caused the committees for epochs to not be correct.

How Has This Been Tested?
---
Existing tests, and manual testing

Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

BREAKING CHANGE: Data directory must be cleaned