feat(tari_swarm_daemon): link up wallet daemon and signaling server (#1041)
Description
---
Links up wallet daemon and signaling server in the `tari_swam`
application

Motivation and Context
---
`tari_swarm_daemon` is a very useful tool to spawn a local Tari network
for development purposes. But currently the wallet daemon process is not
spawn with the signaling server address, which means that webRTC
communication between web pages and the wallet daemon is not possible.
This PR makes the necessary changes to the `tari_swarm_daemon` to make
this possible.

How Has This Been Tested?
---
Manually by running `tari_swarm_daemon` and connecting a site to the
wallet daemon using `tari.js`

What process can a PR reviewer use to test or verify this change?
---
See previous section. Note that existing `tari_swarm` process
directories (and config files) must be created fresh for this change to
work.

Breaking Changes
---

- [x] None
- [ ] Requires data directory to be deleted
- [ ] Other - Please specify

---------

Co-authored-by: stringhandler <mikethetike@tari.com>